+++ _dev/js/theme/core/cart/handler/cart/addToCartHandler.js
@@ -1,5 +1,5 @@
 import useAlertToast from '../../../../components/useAlertToast';
-import addToCartRequest from '../../request/addToCartRequest';
+import addToCartRequest from '../../request/cart/addToCartRequest';
 import sprintf from '../../../../utils/sprintf';
 
 const { danger } = useAlertToast();
