+++ _dev/js/theme/core/cart/request/cart/quantityChangeRequest.js
@@ -1,0 +1,109 @@
+import prestashop from 'prestashop';
+import useDefaultHttpRequest from '../../../../components/http/useDefaultHttpRequest';
+import useHttpPayloadDefinition from '../../../../components/http/useHttpPayloadDefinition';
+
+/**
+ * @typedef ServerResponse
+ * @type {object}
+ * @property {string|string[]} errors - the errors returned by the server
+ * @property {int} id_product - product id
+ * @property {int} id_product_attribute - product attribute id
+ * @property {int} id_customization - product customization id
+ * @property {int} quantity - product quantity
+ * @property {boolean} success - success flag
+ * @property {object} cart - cart front object
+ */
+
+/**
+ * Add voucher to cart request
+ * @param payload {Object} - payload object to send
+ * @param payload.id_product {int} - product id - Required
+ * @param payload.qty {int} - product quantity - Required
+ * @param payload.id_product_attribute {int} - product id attribute - optional pass 0 if not set
+ * @param payload.id_customization {int} - customization id - optional pass 0 if not set
+ * @param payload.add {int} - optional
+ * @param payload.action {string} - optional
+ * @param payload.token {string} - optional
+ * @param payload.ajax {int} - optional
+ * @example
+ *  const payload = {
+ *    id_product: 1, // Required
+ *    qty: 1, // Required
+ *    op: 'up', // up|down Required
+ *    id_product_attribute: 2, // optional
+ *    id_customization: 3, // optional
+ *  };
+ *
+ *  const { getRequest } = quantityChangeRequest(payload);
+ *
+ *  try {
+ *    const resp = await getRequest();
+ *  } catch (error) {
+ *    console.error(error);
+ *  }
+ * @returns {{getRequest: (function(): Promise<ServerResponse>)}}
+ */
+const quantityChangeRequest = (payload) => {
+  const payloadToSend = {
+    update: 1,
+    action: 'update',
+    ajax: 1,
+    token: prestashop.static_token,
+    ...payload,
+  };
+
+  const payloadDefinition = {
+    id_product: {
+      type: 'int',
+      required: true,
+    },
+    qty: {
+      type: 'int',
+      required: true,
+    },
+    id_product_attribute: {
+      type: 'int',
+      required: false,
+    },
+    id_customization: {
+      type: 'int',
+      required: false,
+    },
+    update: {
+      type: 'int',
+      required: true,
+    },
+    action: {
+      type: 'string',
+      required: true,
+    },
+    ajax: {
+      type: 'int',
+      required: true,
+    },
+    op: {
+      type: 'string',
+      required: true,
+    },
+    token: {
+      type: 'string',
+      required: true,
+    },
+  };
+
+  const { validatePayload } = useHttpPayloadDefinition(payloadToSend, payloadDefinition);
+
+  const validationErrors = validatePayload();
+
+  if (validationErrors.length) {
+    throw Error(validationErrors.join(',\n'));
+  }
+
+  const getRequest = () => useDefaultHttpRequest(prestashop.urls.pages.cart, payloadToSend);
+
+  return {
+    getRequest,
+  };
+};
+
+export default quantityChangeRequest;
